agent/requirements.txt
exa-py>=1.0.0
requests>=2.28.0
python-dotenv>=1.0.0
temporalio>=1.5.0
lxml>=4.9.0
